Move general functions scalloc, sstrdup out of i3status.c
