+++ include/i3status.h
@@ -183,13 +183,14 @@ struct min_width {
     const char *str;
 };
 
-char *sstrdup(const char *str);
-
 /* src/general.c */
 char *skip_character(char *input, char character, int amount);
 
 void die(const char *fmt, ...) __attribute__((format(printf, 1, 2), noreturn));
 bool slurp(const char *filename, char *destination, int size);
+char *resolve_tilde(const char *path);
+void *scalloc(size_t size);
+char *sstrdup(const char *str);
 
 /* src/output.c */
 void print_separator(const char *separator);
